+++ lib/log_cleaner/config/routes.rb
@@ -1,4 +1,5 @@
-# log_cleaner/config/routes.rb
+# frozen_string_literal: true
+
 LogCleaner::Engine.routes.draw do
   root to: "dashboard#index"
 end