+++ lib/log_cleaner/logger.rb
@@ -1,11 +1,15 @@
+# frozen_string_literal: true
+
 require "json"
 require "securerandom"
 require "logger"
-require "time"  # Needed for iso8601 timestamps
+require "time" # Needed for iso8601 timestamps
 require_relative "request_store"
 require_relative "config"
 
+# LogCleaner
 module LogCleaner
+  # logger
   class Logger
     def initialize
       # Standard Ruby Logger, output to stdout
@@ -15,10 +19,21 @@ module LogCleaner
     end
 
     # Public methods for log levels
-    def info(data); log("info", data); end
-    def debug(data); log("debug", data); end
-    def warn(data); log("warn", data); end
-    def error(data); log("error", data); end
+    def info(data)
+      log("info", data)
+    end
+
+    def debug(data)
+      log("debug", data)
+    end
+
+    def warn(data)
+      log("warn", data)
+    end
+
+    def error(data)
+      log("error", data)
+    end
 
     private
 
@@ -27,7 +42,7 @@ module LogCleaner
       payload = build_payload(level, data)
       json_pretty = JSON.pretty_generate(payload)
 
-      formatted_msg = "\n#{'*' * 50}\n#{json_pretty}\n#{'*' * 50}\n"
+      formatted_msg = "\n#{"*" * 50}\n#{json_pretty}\n#{"*" * 50}\n"
 
       @logger.public_send(level, formatted_msg)
     end
@@ -39,7 +54,7 @@ module LogCleaner
       {
         timestamp: Time.now.utc.iso8601,
         level: level,
-        request_id: RequestStore.request_id || generate_request_id
+        request_id: RequestStore.request_id || "req-#{SecureRandom.hex(4)}"
       }.merge(filtered_data)
     end
 
@@ -47,9 +62,7 @@ module LogCleaner
     def mask_sensitive(data)
       case data
       when Hash
-        data.transform_keys(&:to_sym).transform_values do |v|
-          mask_sensitive_value(v, data.keys)
-        end
+        mask_sensitive_for_hash(data)
       when Array
         data.map { |v| mask_sensitive(v) }
       else
@@ -57,23 +70,24 @@ module LogCleaner
       end
     end
 
-    private
-
-    # Mask a value if its key is sensitive; recurse otherwise
-    def mask_sensitive_value(value, keys = [])
-      key_sym = keys.first.to_sym rescue nil
-
-      if key_sym && LogCleaner.config.mask_fields.include?(key_sym)
-        "[FILTERED]"
-      else
-        mask_sensitive(value) # recursive call for nested hash/array
+    def mask_sensitive_for_hash(data)
+      data.each_with_object({}) do |(key, value), result|
+        key_sym = safe_to_sym(key)
+        result[key_sym] = mask_sensitive_fileds?(key_sym, value)
       end
     end
 
-    # Generate a request ID if one doesn't exist
-    def generate_request_id
-      "req-#{SecureRandom.hex(4)}"
+    def mask_sensitive_fileds?(key_sym, value)
+      LogCleaner.config.mask_fields.include?(key_sym) ? "[FILTERED]" : mask_sensitive(value)
     end
+
+    # Safely convert a key to symbol
+    def safe_to_sym(key)
+      key.to_sym
+    rescue StandardError
+      key
+    end
+    private :safe_to_sym
   end
 
   # Singleton logger instance
@@ -82,8 +96,19 @@ module LogCleaner
   end
 
   # Convenience methods
-  def self.info(data); logger.info(data); end
-  def self.debug(data); logger.debug(data); end
-  def self.warn(data); logger.warn(data); end
-  def self.error(data); logger.error(data); end
+  def self.info(data)
+    logger.info(data)
+  end
+
+  def self.debug(data)
+    logger.debug(data)
+  end
+
+  def self.warn(data)
+    logger.warn(data)
+  end
+
+  def self.error(data)
+    logger.error(data)
+  end
 end